refactor(web): relocate components out of sidebar dir, chore: update imports
