+++ apps/web/.storybook/decorators/index.ts
@@ -1,3 +1,1 @@
-export { LayoutDecorator, withLayout } from './LayoutDecorator'
-export type { LayoutDecoratorProps } from './LayoutDecorator'
 export { MockProviderDecorator, withMockProvider } from './MockProviderDecorator'
