+++ apps/web/src/components/common/NavTabs/navItemsConfig.tsx
