+++ apps/web/src/components/common/QrCodeButton/index.tsx
