+++ apps/web/src/components/dashboard/PendingTxs/PendingTxsList.tsx
@@ -16,7 +16,7 @@ import useSafeInfo from '@/hooks/useSafeInfo'
 import { useRecoveryQueue } from '@/features/recovery'
 import type { SafeState } from '@safe-global/store/gateway/AUTO_GENERATED/safes'
 import type { RecoveryQueueItem } from '@/features/recovery'
-import { SidebarListItemCounter } from '@/components/sidebar/SidebarList'
+import { SidebarListItemCounter } from '@/components/common/SidebarList'
 
 const PendingRecoveryListItem = dynamic(() => import('./PendingRecoveryListItem'))
