+++ apps/web/src/features/myAccounts/components/MyAccountsV2/components/AccountItem/MultiAccountItem.tsx
@@ -3,7 +3,7 @@ import NextLink from 'next/link'
 import Identicon from '@/components/common/Identicon'
 import AddressWithCopy from '@/components/common/AddressWithCopy'
 import { FiatBalance } from '@/features/spaces'
-import MultiAccountContextMenu from '@/components/sidebar/SafeListContextMenu/MultiAccountContextMenu'
+import MultiAccountContextMenu from '@/features/myAccounts/components/SafeListContextMenu/MultiAccountContextMenu'
 import { AccountItem as BaseAccountItem } from '../../../AccountItem'
 import { AddNetworkButton } from '../../../AddNetworkButton'
 import { useMultiAccountItemData } from '../../../../hooks/useMultiAccountItemData'
